Add a request_received hook
This hook is meant to act like the RequestReceived hook in the canonical
Go example:
https://github.com/twitchtv/twirp/blob/ad6d075411e462f08e971d7efc18f4a770ac8b40/hooks.go#L36-L38

It's not clear to me if the request_received hooks should be called
somewhere within `route_request` or not. For now I'm calling the hooks
before the call to `route_request`, but that might not be what we want?
I don't have much experience with Go so I'm struggling to decipher the
behavior in the canonical implementation.