Add a strict version of JSON encoding.
In some cases when debugging it's nice to have twirp reject JSON
requests that have unknown fields.  This adds a new JSON_STRICT encoding
that will enable this behavior.  It can be opted into by specifying the
strict parameter when specifying the content type.  For curl this looks
like:

```
curl --request POST \
  --url 'http://localhost:8080/twirp/example.hello_world.HelloWorld/Hello' \
  --header 'Content-Type: application/json; strict=true' \
  --data '{"name": "World", "bob": "fail"}'
```