Use the ruby_package when specified
This changes code generation to prefer the ruby_package option when set
in a proto file rather than trying to infer the name based on the
protobuf package name.

We have a case where we use an acronym (DMS for Dead Man's Snitch) in
our packages and they would be generated at `Dms` instead of `DMS` which
we overcome in other code specifying the ruby_package.